feat: update vehicle type route, enhance role-based access control, and refine project configuration
